Added About Us Page and How It Works Page with Team Section
